+++ packages/theme/src/components/TimelinePage.vue
@@ -1,0 +1,7 @@
+<template>
+  <div>时间线页面</div>
+</template>
+
+<script lang="ts" setup></script>
+
+<style lang="scss"></style>