+++ packages/theme/demo/changelog.md
@@ -7,6 +7,16 @@ tag:
 
 # Changelog
 
+## 0.1.9 (2023/02/24)
+### Patch Changes
+- feat: 支持自定义推荐文章的展示顺序，详见[frontmatter.recommend](./config/frontmatter.md#recommend)
+
+
+## 0.1.8 (2023/02/22)
+### Patch Changes
+
+- feat: 支持首页文章置顶能力 (详见[frontmatter.top](./config/frontmatter.md#top)样式较粗糙，欢迎PR改进)
+
 ## 0.1.7 (2023/02/21)
 
 ### Patch Changes